projects/interface/pyproject.toml
[tool.poetry]
name = "interface"
version = "0.1.5"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
tensorflow-macos = "^2.11.0"
scikit-learn = "^1.2.2"
matplotlib = "^3.7.1"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
